requirements.txt
mysql-connector-python
python-dotenv
fastapi
uvicorn[standard]
requests
spyne
lxml
zeep
grpcio
grpcio-tools
protobuf
pysyncobj